IQA-56: Move UserModels to User Subfolder
