+++ web-app/src/models/user/User.py
@@ -1,4 +1,4 @@
-from .. import db
+from ... import db
 from flask_login import UserMixin
 
 class User(UserMixin, db.Model):
