+++ web-app/src/models/user/UserRole.py
@@ -1,4 +1,4 @@
-from .. import db
+from ... import db
 
 class UserRole(db.Model):
